backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings==2.8.0
email-validator==2.2.0
python-multipart==0.0.20
